+++ src/enums/Actions.ts
@@ -80,6 +80,12 @@ export const enum Actions {
   showCreateDialogModal = "show-create-dialog-modal",
 
   /**
+   * Calls a function to update a modal window for creating certain item (file, folder, etc.).
+   * It does not work if the "createDialogProps" parameter is not passed to the message.
+   */
+  updateCreateDialogModal = "update-create-dialog-modal",
+
+  /**
    * Calls a function to open a modal window.
    * It does not work if the "modalDialogProps" parameter is not passed to the message.
    */