+++ src/interfaces/items/IContextMenuItem.ts
@@ -20,6 +20,7 @@ import {
   Devices,
   FilesExst,
   FilesType,
+  FilesSecurity,
   Security,
   UsersType,
 } from "../../enums";
@@ -171,4 +172,12 @@ export interface IContextMenuItem {
    *
    */
   security?: Security[];
+
+  /**
+   * The security parameters of the file or folder or room that will be checked.
+   * If all the parameters are true, the current item will be displayed in the context menu.
+   * If this parameter is undefined, it will be ignored.
+   *
+   */
+  itemSecurity?: (FilesSecurity | Security)[];
 }