+++ src/interfaces/items/IContextMenuItem.ts
@@ -128,6 +128,13 @@ export interface IContextMenuItem {
   onClick: (id: number) => Promise<IMessage> | IMessage | void;
 
   /**
+   * Specifies elements as submenus.
+   * If specified, onClick on the parent will not work.
+   * If none of the child elements are displayed, for example due to security or itemSecurity, the parent will also be hidden.
+   */
+  items?: IContextMenuItem[];
+
+  /**
    * Whether to add the action state to the item in the file list when the onClick event is triggered
    *
    */