+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## Added
 
+- Add withoutBodyPadding and withoutHeaderMargin properties to IModalDialog
+
+## #.#.#
+
+## Added
+
 - Add updateCreateDialogModal action
 - Add isAutoFocusOnError, errorText, onError, onChange properties to ICreateDialog